Paper ref. DFS maze solver
I decided to develop a program to solve a maze then use LaTex to write a paper about what I learned.  I didn’t know at the time that this subject has been done to death on the Web and YouTube.  This paper is not intended to be a real research paper but as a form of geeky play. I get to relearn some things I’ve forgotten, practice programming, and play with the LaTeX software.